// src/modules/inventory/reports/grey-store/grey-fabric-transfer-report/grey-fabric-transfer-report-type.ts
export type IGreyFabricTransferReport = {
  TRANSFER_NO: string;
  TRANSFER_DATE?: Date;
  FROM_BUYER_NAME: string;
  FROM_STYLE_NO: string;
  FROM_PO_NO: string;
  FROM_FABRIC_NAME: string;
  FROM_COLOR_NAME: string;
  FROM_UOM: string;
  FROM_ROLL_QTY: number;
  FROM_QTY_KG: number;
  FROM_GSM: string;
  FROM_DIA: string;
  FROM_STITCH_LENGTH: string;
  TO_BUYER_NAME: string;
  TO_STYLE_NO: string;
  TO_PO_NO: string;
  TO_FABRIC_NAME: string;
  TO_COLOR_NAME: string;
  TO_UOM: string;
  TO_ROLL_QTY: number;
  TO_QTY_KG: number;
  TO_GSM: string;
  TO_DIA: string;
  TO_STITCH_LENGTH: string;
  TO_YARN_CHALLAN_NO: string;
  TO_YARN_LOT: string;
  TO_YARN_BRAND: string;
  TO_YARN: string;
  TO_GREY_SHAPE: string;
  FROM_YARN_CHALLAN_NO: string;
  FROM_YARN_LOT: string;
  FROM_YARN_BRAND: string;
  FROM_YARN: string;
  FROMO_GREY_SHAPE: string;
  CREATED_BY: string;
};